[ApplePushNotification] move functions here
